Commands are now loaded from a path relative to the script (#50)
Prior to this change, it was possible for the commands to go missing if
you invoked the bot from a working dir other than the directory it was
compiled to, e.g `node bin/sweeper.js`. This change makes it so that
the command lookup path is loaded relative to the script, meaning that
the working directory is no longer relevant to whether the commands are
loaded.